Minor fix.
git-svn-id: svn+ssh://svn.code.sf.net/p/itpp/svn/itpp/trunk@644 f17a810d-92cc-4fca-aeb8-6e32867107aa

+++ itpp/base/elmatfunc.h
@@ -195,7 +195,7 @@ namespace itpp {
   double binom(int n, int k);
 
   //! Compute the binomial coefficient "n over k".
-  int binom_i(int n, int k);
+  long binom_i(int n, int k);
 
   //! Compute the base 10 logarithm of the binomial coefficient "n over k".
   double log_binom(int n, int k);
